@@ -75,88 +75,99 @@
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "13.2.6",
-    "@angular/cdk": "13.2.6",
-    "@angular/common": "13.2.6",
-    "@angular/compiler": "13.2.6",
-    "@angular/core": "13.2.6",
-    "@angular/fire": "^7.2.0",
-    "@angular/forms": "13.2.6",
-    "@angular/material": "13.2.6",
-    "@angular/platform-browser": "13.2.6",
-    "@angular/platform-browser-dynamic": "13.2.6",
-    "@angular/router": "13.2.6",
-    "@ngneat/transloco": "3.1.3",
-    "@ngneat/transloco-locale": "^3.0.1",
-    "@ngrx/data": "13.0.2",
-    "@ngrx/effects": "13.0.2",
-    "@ngrx/entity": "13.0.2",
-    "@ngrx/store": "13.0.2",
-    "@ngrx/store-devtools": "13.0.2",
-    "d3": "7.3.0",
-    "firebase": "^9.4.0",
-    "ol": "6.12.0",
-    "rxfire": "^6.0.0",
-    "rxjs": "~6.6.0",
-    "tslib": "^2.0.0",
-    "zone.js": "0.11.5"
+    "@angular/animations": "16.1.8",
+    "@angular/cdk": "16.1.7",
+    "@angular/common": "16.1.8",
+    "@angular/compiler": "16.1.8",
+    "@angular/core": "16.1.8",
+    "@angular/fire": "^7.6.1",
+    "@angular/forms": "16.1.8",
+    "@angular/material": "16.1.7",
+    "@angular/platform-browser": "16.1.8",
+    "@angular/platform-browser-dynamic": "16.1.8",
+    "@angular/router": "16.1.8",
+    "@ngneat/transloco": "5.0.4",
+    "@ngneat/transloco-locale": "^5.0.0",
+    "@ngrx/data": "16.1.0",
+    "@ngrx/effects": "16.1.0",
+    "@ngrx/entity": "16.1.0",
+    "@ngrx/store": "16.1.0",
+    "@ngrx/store-devtools": "16.1.0",
+    "@storybook/core-server": "7.2.1",
+    "@swc/helpers": "~0.5.1",
+    "d3": "7.8.5",
+    "firebase": "10.1.0",
+    "ol": "7.4.0",
+    "rxjs": "7.8.1",
+    "tslib": "^2.6.1",
+    "zone.js": "0.13.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "13.2.6",
-    "@angular-eslint/eslint-plugin": "13.0.1",
-    "@angular-eslint/eslint-plugin-template": "13.0.1",
-    "@angular-eslint/template-parser": "13.0.1",
-    "@angular/cli": "13.2.6",
-    "@angular/compiler-cli": "13.2.6",
-    "@angular/language-service": "13.2.6",
-    "@commitlint/cli": "11.0.0",
-    "@commitlint/config-conventional": "11.0.0",
+    "@angular-devkit/build-angular": "16.1.7",
+    "@angular-devkit/core": "16.1.7",
+    "@angular-devkit/schematics": "16.1.7",
+    "@angular-eslint/eslint-plugin": "16.1.0",
+    "@angular-eslint/eslint-plugin-template": "16.1.0",
+    "@angular-eslint/template-parser": "16.1.0",
+    "@angular/cli": "~16.1.7",
+    "@angular/compiler-cli": "16.1.8",
+    "@angular/language-service": "16.1.8",
+    "@commitlint/cli": "17.6.7",
+    "@commitlint/config-conventional": "17.6.7",
     "@ngneat/tailwind": "^7.0.3",
-    "@nrwl/angular": "13.8.5",
-    "@nrwl/cli": "13.8.5",
-    "@nrwl/cypress": "13.8.5",
-    "@nrwl/devkit": "13.8.5",
-    "@nrwl/eslint-plugin-nx": "13.8.5",
-    "@nrwl/jest": "13.8.5",
-    "@nrwl/linter": "13.8.5",
-    "@nrwl/node": "13.8.5",
-    "@nrwl/storybook": "13.8.5",
-    "@nrwl/tao": "13.8.5",
-    "@nrwl/workspace": "13.8.5",
-    "@storybook/addon-essentials": "6.4.19",
-    "@storybook/angular": "6.4.19",
-    "@storybook/builder-webpack5": "6.4.19",
-    "@storybook/manager-webpack5": "6.4.19",
-    "@types/d3": "7.0.0",
-    "@types/jest": "26.0.24",
-    "@types/node": "14.14.33",
-    "@typescript-eslint/eslint-plugin": "5.10.2",
-    "@typescript-eslint/parser": "5.10.2",
-    "autoprefixer": "^10.3.1",
-    "cypress": "^7.3.0",
-    "dotenv": "~10.0.0",
-    "eslint": "8.7.0",
-    "eslint-config-prettier": "8.1.0",
-    "eslint-plugin-cypress": "^2.10.3",
-    "eslint-plugin-deprecation": "1.2.1",
-    "eslint-plugin-import": "2.22.1",
-    "eslint-plugin-jsdoc": "32.2.0",
+    "@nx/angular": "16.6.0",
+    "@nx/cypress": "16.6.0",
+    "@nx/devkit": "16.6.0",
+    "@nx/eslint-plugin": "16.6.0",
+    "@nx/jest": "16.6.0",
+    "@nx/js": "16.6.0",
+    "@nx/linter": "16.6.0",
+    "@nx/node": "16.6.0",
+    "@nx/plugin": "16.6.0",
+    "@nx/storybook": "16.6.0",
+    "@nx/workspace": "16.6.0",
+    "@schematics/angular": "16.1.7",
+    "@storybook/addon-essentials": "7.2.1",
+    "@storybook/angular": "7.2.1",
+    "@storybook/builder-webpack5": "7.2.1",
+    "@storybook/manager-webpack5": "6.5.16",
+    "@swc-node/register": "~1.6.6",
+    "@swc/cli": "~0.1.62",
+    "@swc/core": "~1.3.74",
+    "@types/d3": "7.4.0",
+    "@types/jest": "29.5.3",
+    "@types/node": "20.4.7",
+    "@typescript-eslint/eslint-plugin": "5.60.1",
+    "@typescript-eslint/parser": "5.60.1",
+    "autoprefixer": "^10.4.14",
+    "cypress": "12.17.3",
+    "dotenv": "~16.3.1",
+    "eslint": "8.46.0",
+    "eslint-config-prettier": "8.10.0",
+    "eslint-plugin-cypress": "^2.13.3",
+    "eslint-plugin-deprecation": "1.5.0",
+    "eslint-plugin-import": "2.28.0",
+    "eslint-plugin-jsdoc": "46.4.5",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prefer-arrow": "1.2.3",
-    "eslint-plugin-promise": "^6.0.0",
-    "eslint-plugin-standard": "^5.0.0",
-    "husky": "4.3.8",
-    "jest": "27.0.3",
-    "jest-preset-angular": "11.1.1",
-    "ng-packagr": "13.2.1",
-    "postcss": "^8.3.6",
-    "postcss-import": "^14.0.2",
-    "postcss-preset-env": "^6.7.0",
-    "postcss-url": "^10.1.2",
-    "prettier": "2.5.1",
-    "tailwindcss": "^2.2.7",
-    "ts-jest": "27.0.3",
-    "ts-node": "~9.1.1",
-    "typescript": "4.5.5"
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-config-standard": "^17.1.0",
+    "husky": "8.0.3",
+    "jest": "29.6.2",
+    "jest-environment-jsdom": "^29.6.2",
+    "jest-preset-angular": "13.1.1",
+    "ng-packagr": "16.1.0",
+    "nx": "16.6.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "postcss": "^8.4.27",
+    "postcss-import": "15.1.0",
+    "postcss-preset-env": "9.1.0",
+    "postcss-url": "10.1.3",
+    "prettier": "3.0.1",
+    "tailwindcss": "3.3.3",
+    "ts-jest": "29.1.1",
+    "ts-node": "10.9.1",
+    "typescript": "5.1.6"
   }
 }
