@@ -1,11 +1,7 @@
-const rootMain = require('../../../.storybook/main');
 
-rootMain.core = {...rootMain.core, builder: 'webpack5'};
-
-// Use the following syntax to add addons!
-// rootMain.addons.push('');
-rootMain.stories.push(
-  ...['../**/*.stories.@(js|jsx|ts|tsx)', '../**/**/*.stories.@(js|jsx|ts|tsx)']
-);
-
-module.exports = rootMain;
+module.exports = {
+  core: { builder: 'webpack5' },
+  stories: ['../**/*.stories.@(js|jsx|ts|tsx)',
+    '../**/**/*.stories.@(js|jsx|ts|tsx)',],
+  addons: ['@storybook/addon-essentials'],
+};
