@@ -1,39 +1,44 @@
-import { chain, Rule, SchematicContext, Tree } from '@angular-devkit/schematics';
-import { execSync } from 'child_process';
+import {chain, Rule, SchematicContext, Tree} from '@angular-devkit/schematics';
+import {execSync} from 'child_process';
 
 const libs = ['util', 'ui', 'feature', 'data-access'];
 
-
 const nxClearCache = (): Rule => {
-  execSync('nx clear-cache',{stdio: 'inherit'}); 
+  execSync('nx clear-cache', {stdio: 'inherit'});
   return () => {};
-}
-const lintLibs = (): Rule => { 
-  execSync('nx workspace-lint && nx run-many --projects=data-access,feature,ui,util --target=lint',{stdio: 'inherit'}); 
+};
+const lintLibs = (): Rule => {
+  execSync(
+    'nx workspace-lint && nx run-many --projects=data-access,feature,ui,util --target=lint',
+    {stdio: 'inherit'}
+  );
   return () => {};
-}
-const testLibs = (): Rule =>  { 
-  execSync('nx run-many --projects=data-access,feature,ui,util --target=test --browsers ChromeHeadless',{stdio: 'inherit'}); 
+};
+const testLibs = (): Rule => {
+  execSync(
+    'nx run-many --projects=data-access,feature,ui,util --target=test --browsers ChromeHeadless',
+    {stdio: 'inherit'}
+  );
   return () => {};
-}
-const buildLibs = (): Rule =>  { 
-  execSync('nx run-many --projects=data-access,feature,ui,util --target=build',{stdio: 'inherit'});
+};
+const buildLibs = (): Rule => {
+  execSync('nx run-many --projects=data-access,feature,ui,util --target=build', {stdio: 'inherit'});
   return () => {};
-}
-const publishLibs = (version: string): Rule => {   
+};
+const publishLibs = (version: string): Rule => {
   libs.forEach((lib: string) => {
     let command = '';
     command += 'cd dist/libs/' + lib;
     command += '&&';
-    command += `npm version ${version}`;
-    command += '&&';
+    //command += `npm version ${version}`;
+    //command += '&&';
     command += 'npm pack';
     command += '&&';
     command += 'npm publish';
-    execSync(command, { stdio: 'inherit' });
+    execSync(command, {stdio: 'inherit'});
   });
   return () => {};
-}
+};
 
 export default function (tree: Tree, schema: Record<string, never>) {
   const version = String(schema.version);
@@ -41,9 +46,8 @@ export default function (tree: Tree, schema: Record<string, never>) {
     return;
   }
   console.log('Verion: ', version);
-  return chain([nxClearCache(),lintLibs(),testLibs(),buildLibs(), publishLibs(version)]);
+  return chain([nxClearCache(), lintLibs(), testLibs(), buildLibs(), publishLibs(version)]);
 }
-
 
 /*
     "npm_publish_util": "cd dist/libs/util && npm version 0.1.2 && npm pack && npm publish",
